apps/ai/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
openai==1.101.0
pillow==11.3.0
numpy==2.2.6
python-multipart==0.0.20
pydantic==2.11.7
httpx==0.28.1
